feat(model): Working on polymorphism and stream method name

+++ module/model-dynamodb/src/service.ts
@@ -4,7 +4,7 @@ import { Class, ShutdownManager, Util } from '@travetto/base';
 import { Injectable } from '@travetto/di';
 import {
   ModelCrudSupport, ModelExpirySupport, ModelRegistry, ModelStorageSupport,
-  ModelIndexedSupport, ModelType, NotFoundError, ExistsError
+  ModelIndexedSupport, ModelType, NotFoundError, ExistsError, SubTypeNotSupportedError
 } from '@travetto/model';
 
 import { ModelCrudUtil } from '@travetto/model/src/internal/service/crud';
@@ -245,24 +245,40 @@ export class DynamoDBModelService implements ModelCrudSupport, ModelExpirySuppor
   }
 
   async update<T extends ModelType>(cls: Class<T>, item: T) {
+    if (ModelRegistry.get(cls).subType) {
+      throw new SubTypeNotSupportedError(cls);
+    }
     item = await ModelCrudUtil.preStore(cls, item, this);
+    if (ModelRegistry.get(cls).expiresAt) {
+      await this.get(cls, item.id);
+    }
     await this.putItem(cls, item.id, item, 'update');
     return item;
   }
 
   async upsert<T extends ModelType>(cls: Class<T>, item: T) {
+    if (ModelRegistry.get(cls).subType) {
+      throw new SubTypeNotSupportedError(cls);
+    }
     item = await ModelCrudUtil.preStore(cls, item, this);
     await this.putItem(cls, item.id, item, 'upsert');
     return item;
   }
 
-  async updatePartial<T extends ModelType>(cls: Class<T>, id: string, item: Partial<T>, view?: string) {
+  async updatePartial<T extends ModelType>(cls: Class<T>, item: Partial<T> & { id: string }, view?: string) {
+    if (ModelRegistry.get(cls).subType) {
+      throw new SubTypeNotSupportedError(cls);
+    }
+    const id = item.id;
     item = await ModelCrudUtil.naivePartialUpdate(cls, item, view, () => this.get(cls, id)) as T;
     await this.putItem(cls, id, item as T, 'update');
     return item as T;
   }
 
   async delete<T extends ModelType>(cls: Class<T>, id: string) {
+    if (ModelRegistry.get(cls).subType) {
+      throw new SubTypeNotSupportedError(cls);
+    }
     const res = await this.cl.deleteItem({
       TableName: this.resolveTable(cls),
       ReturnValues: 'ALL_OLD',
@@ -285,7 +301,7 @@ export class DynamoDBModelService implements ModelCrudSupport, ModelExpirySuppor
       if (batch.Count && batch.Items) {
         for (const el of batch.Items) {
           try {
-            yield loadAndCheckExpiry(cls, el.body.S!);
+            yield await loadAndCheckExpiry(cls, el.body.S!);
           } catch (e) {
             if (!(e instanceof NotFoundError)) {
               throw e;
@@ -309,6 +325,10 @@ export class DynamoDBModelService implements ModelCrudSupport, ModelExpirySuppor
 
   // Indexed
   async getByIndex<T extends ModelType>(cls: Class<T>, idx: string, body: Partial<T>) {
+    if (ModelRegistry.get(cls).subType) {
+      throw new SubTypeNotSupportedError(cls);
+    }
+
     const query = {
       TableName: this.resolveTable(cls),
       IndexName: idx,
@@ -328,6 +348,10 @@ export class DynamoDBModelService implements ModelCrudSupport, ModelExpirySuppor
   }
 
   async deleteByIndex<T extends ModelType>(cls: Class<T>, idx: string, body: Partial<T>) {
+    if (ModelRegistry.get(cls).subType) {
+      throw new SubTypeNotSupportedError(cls);
+    }
+
     const query = {
       TableName: this.resolveTable(cls),
       IndexName: idx,
